update order, active filds

+++ database/migrations/2022_06_27_155157_create_tables_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->uuid('uuid')->unique();
             $table->unsignedInteger('project_id')->nullable()->default(NULL);
             $table->string('name')->nullable()->default(NULL);
-            $table->smallInteger('active')->nullable()->default(NULL);
-            $table->tinyInteger('order')->nullable()->default(NULL);
+            $table->tinyInteger('active')->nullable()->default(NULL);
+            $table->smallInteger('order')->nullable()->default(NULL);
             
             $table->timestamps();
         });
