+++ database/migrations/2022_06_27_181338_create_projects_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('name')->nullable()->default(NULL);
             $table->string('alias')->unique();
             $table->string('logo')->nullable()->default(NULL);
-            $table->smallInteger('active')->nullable()->default(NULL);
+            $table->tinyInteger('active')->nullable()->default(NULL);
 
             $table->timestamps();
         });
